Making it possible to edit any field in coreData, not just numeric ones. Also making sure that input values containing timezones actually contain them.
